backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
openai>=1.50.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
faiss-cpu>=1.7.0
numpy>=1.26.0
